This is a code complete commit. Application testing still needs to be done and more unit tests need to be written. Also need to complete Exactly Once implementation of DimensionStore and aggregation specification in dimensional computation.
Updated Validators and Deserializers to take a context object.
Allowed the user to set context objects on dataDeserializerFactories.
Fixed bugs and updated GenericSchemaClasses.
Updated the GenericDataQuery class to meet latest spec.
Created a new implementation of the QueueManager to cause new queries with the same ID to replace existing queries with the same ID.
Implemented Generic Dimension computation operator and Generic Dimension Computation store operators as well as generic aggregators
Create a new version of the AdsDemo To use Generic Dimension Computation operators as well as genericized DataQuery And Results.